Improved the SPARQL query for properties.
The SPARQL query for properties required a direct ‘rdf:Property’ type
relation, which is often not present in linked datasets.  This commit
also support for properties that have a direct ‘owl:DatatypeProperty’
or ‘owl:ObjectProperty’ type relation.  The combination of these three
alternative characterizations of a property should be able to match
most properties that are exposed in linked data.

This commit also cleans up the class and property queries a bit,
including the removal of superfluous string casts and deduplication of
prefix declarations.